+++ include/dfwimpl/config.h
@@ -36,7 +36,7 @@ class config:
 
 	private:
 
-	std::string get_file_path(const tile_editor::env& _env) const {return _env.build_data_path("config/config.json");}
+	std::string get_file_path(const tile_editor::env& _env) const {return _env.build_user_path("config.json");}
 };
 
 dfw::input_description          input_description_from_config_token(const rapidjson::Value&);