+++ include/tile_editor/app/env.h
@@ -13,7 +13,6 @@ struct env {
 * found (configuration, mostly).
 */
 //	std::string	get_app_path() const;
-	std::string build_log_path(const std::string&) const;
 	std::string build_assets_path(const std::string&) const;
 	std::string build_data_path(const std::string&) const;