+++ make_appimage.sh
@@ -9,10 +9,10 @@ fi;
 #export SIGN=1
 #export SIGN_KEY=0xLAST8CHARSOFGPGKEYID
 
-./build.sh appimage
 $2 --appdir AppDir
-cp -r resources/data AppDir/usr/share
+cp -r resources/* AppDir/usr/share
 mkdir -p AppDir/usr/share/metainfo
 cp linuxdeploy/*.xml AppDir/usr/share/metainfo/
 $2 --executable $1 --appdir AppDir -i ./linuxdeploy/tile_editor.png -d ./linuxdeploy/tile_editor.desktop --output appimage
+rm -rf AppDir
 exit 0