+++ src/tile_editor/env.cpp
@@ -3,17 +3,12 @@
 using namespace tile_editor;
 
 env::env(
-	const std::string& _exec, 
+	const std::string& _exec,
 	const std::string _home
 ):
 	exec_dir(_exec),
 	userdir{_home+"/.tile_editor/"}
 {}
-
-std::string env::build_log_path(const std::string& _file) const {
-
-	return exec_dir+std::string{"logs/"}+_file;
-}
 
 std::string env::build_data_path(const std::string& _file) const {